short script to check file extensions
I've written this short script over the weekend to go through the list of filenames in .xlsx format generated by the main app and return a text file listing all different file extensions. The purpose is that when I will eventually get around to comparing and removing duplicates, I want to check only audio files and exclude other files (e.g. "front.jpg", an image of the front cover of an album, found hundreds if not thousands of times with the same name but depicting a different image)